Implement fix_variable
